Guard for forge flag in default Magit bindings
The Magit module does not automatically install Forge. This means that
when you attempt to use it without that flag enabled, you end up with
some bindings to commands that do not exist.

This change makes it so we do not set those bindings when the Forge flag
isn't set.